exec: add assert_debug_safe and notes on debug structures
The debug structures are different from many of the other shared ones as
they are modified from two places:

  - architectural debug registers
  - the gdb stub

The first is usually in the context of vCPU currently running so
modifications are safe. The second is generally only changed while the
vCPUs are halted and therefore not going to clash.

KNOWN ISSUES: arm powerctl resets will trigger the asserts in
multi-thread mode.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>